Pull Interpreted type into 'LS.Interpreter' module
The Interpreted type is an "interpreted" form of `[Rule]`. As such, it
contains meta-information such as ScopeTables and a rule dependency
graph.

If `Interpreted` lives in `LS.Rule`, then all types that are introduced
for various phases of the transpilation pipeline *must* also be defined
in `LS.Rule` to avoid a cyclic dependency.
There is no good reason for this, so we pull `Interpreted` into the
`LS.Interpreter` module, and untangle cyclic module dependencies we
uncover.

In particular, the `LS.PrettyPrinter` used to import `LS.Interpreted` to
provide some pretty instances for "type signatures", e.g. `ParamText`
and `TypeSig` which which depends on some context in `Interpreted`.
However, `LS.Interpreted` also needs to depend on `LS.PrettyPrinter` as
it needs the utilities to provide logging output.
To fix this, we pull out pretty instances for type signatures into its
own module, allowing us to break the cycle.

As `Interpreted` is pervasively used in the codebase, this refactoring
requires us to touch many transpilers and test cases. However, these
changes are trivial.
