+++ lib/haskell/natural4/src/LS/Interpreter.hs
@@ -22,7 +22,14 @@
 --
 -- To view the various endpoints exposed by the Interpreter, see @org/LATEST.org@.
 module LS.Interpreter
-  ( RuleSet,
+  ( -- * Main Interpreter
+    Interpreted(..),
+    defaultL4I,
+    RuleGraph,
+    RuleGraphEdgeLabel,
+    l4interpret,
+    -- * Interpreter helpers
+    RuleSet,
     allCTkeys,
     attrsAsMethods,
     attrType,
@@ -50,7 +57,6 @@ module LS.Interpreter
     isAnEnum,
     isRuleAlias,
     itemsByRule,
-    l4interpret,
     onlyTheItems,
     qaHornsR,
     qaHornsT,
@@ -58,6 +64,7 @@ module LS.Interpreter
     ruleLocals,
     toObjectStr,
     topsortedClasses,
+    typeOfTerm,
   )
 where
 
@@ -79,6 +86,7 @@ import Data.Graph.Inductive
     nodes,
     topsort,
   )
+import Data.Graph.Inductive qualified as Gr
 import Data.HashMap.Strict qualified as Map
 import Data.List (find, (\\))
 import Data.List qualified as DL
@@ -105,8 +113,7 @@ import LS.RelationalPredicates
     partitionExistentials,
   )
 import LS.Rule
-  ( Interpreted (..),
-    Rule
+  ( Rule
       ( DefTypically,
         Hornlike,
         Regulative,
@@ -137,8 +144,6 @@ import LS.Rule
         who,
         wwhere
       ),
-    RuleGraph,
-    RuleGraphEdgeLabel,
     ValuePredicate
       ( attrCond,
         attrName,
@@ -207,10 +212,68 @@ import LS.XPile.Logging
     pShowNoColorS,
     xpError,
     xpLog,
-    xpReturn,
+    xpReturn, XPileLogW,
   )
 import Text.Pretty.Simple (pShowNoColor)
 import Text.Regex.PCRE.Heavy qualified as PCRE
+
+
+-- | This is generated by the Interpreter and handed around to different transpilers.
+--
+-- In the future we may add to this list of attributes the following frequently called functions:
+--
+-- * `LS.Interpreter.qaHornsT` -- currently a function run against `l4i`
+--
+-- * `LS.Interpreter.qaHornsR` -- currently a function run against `l4i`
+
+data Interpreted = L4I {
+  -- | all the `DECLARE`d classes
+  classtable :: ClsTab
+
+  -- | all the rule scopes; the exact semantics are not
+  -- entirely nailed down as we don't have a rigorous notion of scope
+  -- yet. This is intended to reflect the real-world usage of "for the
+  -- purposes of this section, X means Y". We might think of that as a
+  -- nested block scope for @let@ bindings.
+  , scopetable :: ScopeTabs
+
+  -- | the original rules on which the interpretation was based. You
+  -- may sometimes see functions that take both 'Interpreted' and
+  -- @[Rule]@; the latter is technically redundant and can be safely
+  -- eliminated. [TODO].
+  , origrules  :: [Rule]
+
+  -- | valuepredicates contain the bulk of the top-level decision logic, and can be easily expressed as instance or class methosd.
+  , valuePreds :: [ValuePredicate]
+
+  -- | rule decision graph gets used by multiple transpilers, so it lives here
+  , ruleGraph :: RuleGraph
+  , ruleGraphErr :: XPileLogW
+  }
+  deriving (Eq, Show)
+
+-- | default L4I
+defaultL4I :: Interpreted
+defaultL4I = L4I
+  { classtable = CT Map.empty
+  , scopetable = Map.empty
+  , origrules = mempty
+  , valuePreds = mempty
+  , ruleGraph = Gr.empty
+  , ruleGraphErr = mempty
+  }
+
+-- | structure the rules as a graph.
+-- in the simple case, the graph is one or more trees, each rooted at a "top-level" rule which is not "used" by any another rule.
+--
+-- if we walk the roots, we will sooner or later encounter all the decision elements relevant to each root.
+-- in a less simple case, the graph is cyclic! everything depends on everything else! but we can recognize that as an error condition.
+--
+-- note that a regulative rule R1 HENCE R2 is recorded as a single rule, even if we think of the R2 as a separate rule
+-- perhaps we should have a notion of anonymous rules, that are internally labelled and structured, so R2 is equal to R1 in the graph.
+
+type RuleGraphEdgeLabel = ()
+type RuleGraph = Gr Rule RuleGraphEdgeLabel
 
 -- | interpret the parsed rules based on some configuration options.
 -- This is a canonical intermediate representation used by downstream
@@ -975,7 +1038,7 @@ extractRPMT2Text _         =
 type RuleSet = [Rule]
 
 -- * the getRuleBy* family
--- 
+--
 -- | Retrieve a rule by name, using `ruleName`
 getRuleByName :: RuleSet -> RuleName -> Maybe Rule
 getRuleByName rs rn = find (\r -> ruleName r == rn) rs
@@ -1097,13 +1160,13 @@ globalFacts l4i = do
 --           ClassA's RecordAttr's AttributeName IS baz WHEN quux
 --           ClassA's RecordAttr's AttributeName IS baz OTHERWISE
 -- @
--- 
+--
 -- Extract and return all of these decisions in the form:
 --
 -- @
 --    [([ClassA, RecordAttr, AttributeName], foo, Just bar
 -- @
--- 
+--
 -- | we extract the methods to a fully qualified and annotatable form defined as `ValuePredicate` -- see Types.hs
 --
 -- Go through every rule and break it down into this structure.
@@ -1160,7 +1223,7 @@ attrsAsMethods rs = do
           xpError ["unhandled RelationalPredicate", show hHead]
 
 -- | input: [MTT "foo's", MTT "bar's", MTT "baz"]
--- 
+--
 --  output: (["foo", "bar"], "baz")
 toObjectPath :: MultiTerm -> XPileLogE ([EntityName], EntityName)
 toObjectPath [] = do
@@ -1232,6 +1295,19 @@ enumMatch enumNames _mt (_givenName, Just _mgiven@(SimpleType TOne etype)) =
 lowerMT :: MTExpr -> MTExpr
 lowerMT (MTT t) = MTT $ T.toLower t
 lowerMT x       = x
+
+-- | what is the (L4) type of a given term?
+-- the term in question is given as a TypedMulti, which often works out to be a key/value together in a TypedMulti.
+-- if the TypedMulti already contains an explicit type, we'll return that.
+-- otherwise, we need to interrogate the class hierarchy and the "varpath" breadcrumb by which we arrived at the current term.
+-- for example, if we know that a Corporation extends Party with a representative : Natural Person, and we are now considering
+-- a representative: Bob inside of a const Company : US_Company which extends Corporation, we have enough information to know that
+-- the RHS Bob is a variable name, not a string, so we shoudn't double-quote it.
+typeOfTerm :: Interpreted {- -> VarPath -} -> TypedMulti -> Maybe TypeSig
+typeOfTerm l4i _tm =
+  let (CT _ch, _scopetabs) = (classtable l4i, scopetable l4i)
+  in
+    Nothing
 
 -- actually i think this is not needed because we can use `extendedAttributes`
 -- -- | an expanded version of the class table, respecting inheritance, and respecting overrides along the way.
