+++ lib/haskell/natural4/src/LS/PrettyPrinter.hs
@@ -10,15 +10,11 @@
 -- This is similar to instantiating into Show, but it all happens within Prettyprinter's "Doc ann" rather than String.
 -- The pretty-printing then gets used by the transpilers.
 module LS.PrettyPrinter
-  ( ParamText3 (..),
-    ParamText4 (..),
-    RP1 (..),
+  ( RP1 (..),
     commentWith,
     inPredicateForm,
     myrender,
     orgexample,
-    prettyMaybeType,
-    prettySimpleType,
     snake_case,
     snake_inner,
     srchs,
@@ -27,6 +23,9 @@ module LS.PrettyPrinter
     vvsep,
     (<//>),
     (</>),
+    quoteRHS,
+    -- * Helpers to write more pretty instances!
+    MTQ(..),
   )
 where
 
@@ -35,19 +34,14 @@ import Control.Arrow ((>>>))
 import Data.Coerce (coerce)
 import Data.Foldable qualified as DF
 import Data.List (intersperse)
-import Data.List.NonEmpty as NE (NonEmpty ((:|)), head, tail, toList)
 import Data.String.Interpolate (i)
 import Data.String.Interpolate.Conversion (Interpolatable)
 import Data.Text qualified as T
 import Data.Traversable qualified as DT
 import Debug.Trace (trace)
-import LS.Rule (Interpreted (classtable, scopetable))
 import LS.Types
-  ( ClsTab (..),
-    MTExpr (..),
+  ( MTExpr (..),
     MultiTerm,
-    ParamText,
-    ParamType (TList0, TList1, TOne, TOptional, TSet0, TSet1),
     RPRel (RPhas, RPis),
     RelationalPredicate (..),
     TypeSig (..),
@@ -62,18 +56,12 @@ import Prettyprinter
     LayoutOptions (layoutPageWidth),
     PageWidth (Unbounded),
     Pretty (pretty),
-    brackets,
-    colon,
-    comma,
     defaultLayoutOptions,
     dquotes,
     encloseSep,
-    hcat,
     hsep,
     layoutPretty,
     line,
-    nest,
-    parens,
     vsep,
     (<+>),
   )
@@ -132,7 +120,7 @@ instance Pretty RP1 where
 
 -- | [TODO] this is lossy -- preserve the original cell type information from the underlying multiterm in the relationalpredicate
 inPredicateForm :: RelationalPredicate -> MultiTerm
-inPredicateForm (RPParamText   pt)                = toList . (fst =<<) $ pt
+inPredicateForm (RPParamText   pt)                = DF.toList . (fst =<<) $ pt
 inPredicateForm (RPMT     [MTT "OTHERWISE"])      = mempty
 inPredicateForm (RPMT          mt)                = pred_flip mt
 inPredicateForm (RPConstraint  mt  RPis  [MTT "Yes"]) = pred_flip mt
@@ -245,60 +233,6 @@ untaint =
 instance Pretty RPRel where
   pretty rpr = pretty $ rel2txt rpr
 
-newtype ParamText2 = PT2 ParamText
-
--- | ParamText2: the first line appears on the first line, subsequent lines appear on subsequent lines
--- if you want different ways of rendering a ParamText, go ahead and create newtypes for it.
--- because a ParamText is a type alias.
--- so it makes sense to marshal to a newtype PT2, PT3, PT4 etc that has the appropriate rendering instance.
-instance Pretty ParamText2 where
-  pretty (PT2 (x1 :| xs)) = nest 4 ( vsep ( hsep (pretty <$> NE.toList ( fst x1) )
-                                           : (hsep . fmap pretty . toList . fst <$> xs) ) )
-
--- | ParamText3 is used by the CoreL4 transpiler to produce colon-annotated paramtexts.
-newtype ParamText3 = PT3 ParamText
-instance Pretty ParamText3 where
-  pretty (PT3 pt) = hcat $ intersperse ", " $ toList $ typedOrNot "_" <$> pt
-
--- | ParamText4 is used to approximate a recursive record. currently we can only go 2 deep.
--- in future we will have to upgrade ParamText to a full Tree type which can nest arbitrarily deep.
-data ParamText4 = PT4 ParamText Interpreted -- VarPath
-                | PT5 ParamText Interpreted
-  deriving (Eq, Show)
-
-instance Pretty ParamText4 where
-  pretty (PT5 orig@(line1 :| line2s) l4i)
-    | null line2s = [di|#{word1 line1} = #{quoteBoT l4i line1}|]
-    | otherwise   = [di|#{word1 line1} : #{quoteBoT l4i line1}|]
-    
-  pretty (PT4 orig@(line1 :| line2s) l4i) -- varpath)
-    | null line2s = -- "//" <+> "208:" <+> viaShow line1 <//>
-                     [di|#{quoteRHS line1} : #{pretty (MT1 (NE.head (fst line1)))}|]
-                     -- we should be in a DEFINE, printing a value; if we're not, we may be dumping values with the wrong order, so we need to create a PT5.
-                     --    | line2s == [] = "-- " <> viaShow line1 <//> "199: " <> word1 line1 <> colon <+> quoteBoT line1
-                     
-    | otherwise    = [di|// 213: #{orig}|] <//> -- [TODO] need to fix this -- test by considering a DEFINE with nested records.
-                     [__di|
-                      #{quoteRHS line1} = { -- #{quoteBoT l4i line1}
-                        #{vsep [ word1 l2 <+> colon <+> dquotes (lrest l2) <> comma | l2 <- line2s ]}
-                      }
-                     |]
-
-word1,lrest :: TypedMulti -> Doc ann
-word1 l = typedOrNot "_"      ((NE.head . fst $ l) :| [], snd l)
-lrest l = hsep $ pretty . T.replace "\n" "\\n" <$> (NE.tail . fmap mtexpr2text . fst $ l)
--- quote based on type.
-quoteBoT :: Interpreted -> TypedMulti -> Doc ann
-quoteBoT l4i l@(net, _mts) =
-  let unquoted = hsep $ pretty <$> NE.tail net
-  in case typeOfTerm l4i {-varpath-} l of
-    Just (SimpleType _ s1)  -> case s1 of
-                                 "string" -> dquotes $ lrest l
-                                 "number" -> unquoted
-                                 "num"    -> unquoted
-                                 _        -> snake_case $ NE.tail net
-    Just (InlineEnum _p _s) -> dquotes $ lrest l
-    Nothing                 -> unquoted
 -- quote for a DEFINE value, where we are saying X IS THE Y
 -- "Y" = X;
 quoteRHS :: TypedMulti -> Doc ann
@@ -307,20 +241,6 @@ quoteRHS l@(val, mts) =
     Just (SimpleType _ s1) -> snake_case [MTT s1]
     unexpected             -> trace [i|ERROR: PrettyPrinter PT4 quoteRHS: surprised to see the type annotation #{unexpected}|] $
                               dquotes "ERRORTYPE_PrettyPrinter_quoteRHS"
-
--- what is the (L4) type of a given term?
--- the term in question is given as a TypedMulti, which often works out to be a key/value together in a TypedMulti.
--- if the TypedMulti already contains an explicit type, we'll return that.
--- otherwise, we need to interrogate the class hierarchy and the "varpath" breadcrumb by which we arrived at the current term.
--- for example, if we know that a Corporation extends Party with a representative : Natural Person, and we are now considering
--- a representative: Bob inside of a const Company : US_Company which extends Corporation, we have enough information to know that
--- the RHS Bob is a variable name, not a string, so we shoudn't double-quote it.
-
-typeOfTerm :: Interpreted {- -> VarPath -} -> TypedMulti -> Maybe TypeSig
-typeOfTerm l4i _tm =
-  let (CT _ch, _scopetabs) = (classtable l4i, scopetable l4i)
-  in
-    Nothing
 
 -- -- we are given a long varpath representing the traversal through class (anonymous) attribute space -- i.e. multiple levels of HAS
 -- walk :: Interpreted -> Maybe ClsTab -> VarPath -> TypedMulti -> Maybe TypeSig
@@ -336,7 +256,7 @@ typeOfTerm l4i _tm =
 --                                    Just (its, ct1) -> trace ("underlying type " ++ show t1 ++ " with inferrable typesig " ++ show its ++ " found in toplevel classtable, walking") $
 --                                                       trace ("attempting to obtain extended attributes first tho") $
 --                                                       let ct2 = extendedAttributes ct1 t1
---                                                       in 
+--                                                       in
 --                                                       walk l4i ct2 xs ot
 --                      Left err -> trace ("underlying type returned error: " ++ err) $
 --                                  Nothing
@@ -347,38 +267,6 @@ typeOfTerm l4i _tm =
 --                                 Nothing
 --                      Just (its, ct1) -> trace ("yes it does! we got back ct2 " ++ show ct1 ++ " so we will recurse, dropping " ++ show x) $
 --                                         walk l4i (Just ct1) xs ot
-
-typedOrNot :: String -> TypedMulti -> Doc ann
-typedOrNot        _ (multiterm, Nothing)                        = [di|#{snake_case (toList multiterm)}: Object|]
-typedOrNot        _ (multiterm, Just (SimpleType TOne      s1)) = [di|#{snake_case (toList multiterm)}: #{snake_case [MTT s1]}|]
-typedOrNot "corel4" (multiterm, Just (SimpleType TOptional s1)) = [di|#{snake_case (toList multiterm)}: #{[MTT s1]}|]
-typedOrNot        _ (multiterm, Just (SimpleType TOptional s1)) = [di|#{snake_case (toList multiterm)}:? #{snake_case [MTT s1]}|]
-typedOrNot        _ (multiterm, Just (SimpleType TList0    s1)) = [di|#{snake_case (toList multiterm)}: #{brackets (pretty s1)}|]
-typedOrNot        _ (multiterm, Just (SimpleType TList1    s1)) = [di|#{snake_case (toList multiterm)}: #{brackets (pretty s1)}|]
-typedOrNot        _ (multiterm, Just (SimpleType TSet0     s1)) = [di|#{snake_case (toList multiterm)}: #{brackets (pretty s1)}|]
-typedOrNot        _ (multiterm, Just (SimpleType TSet1     s1)) = [di|#{snake_case (toList multiterm)}: #{brackets (pretty s1)}|]
-typedOrNot        _ (multiterm, Just (InlineEnum pt1       s1)) =
-  [di|#{snake_case (toList multiterm)}\# : InlineEnum unsupported: #{pt1} #{parens (pretty $ PT2 s1)}|]
-
-prettySimpleType :: String -> (T.Text -> Doc ann) -> TypeSig -> Doc ann
-prettySimpleType _        prty (SimpleType TOne      s1) = prty s1
-prettySimpleType "corel4" prty (SimpleType TOptional s1) = prty s1
-prettySimpleType "ts"     prty (SimpleType TOptional s1) = prty s1
-prettySimpleType _        prty (SimpleType TOptional s1) = [di|#{prty s1}?|]
-prettySimpleType "ts"     prty (SimpleType TList0    s1) = [di|#{prty s1}[]|]
-prettySimpleType "ts"     prty (SimpleType TList1    s1) = [di|#{prty s1}[]|]
-prettySimpleType "ts"     prty (SimpleType TSet0     s1) = [di|#{prty s1}[]|]
-prettySimpleType "ts"     prty (SimpleType TSet1     s1) = [di|#{prty s1}[]|]
-prettySimpleType _        prty (SimpleType TList0    s1) = [di|[#{prty s1}]|]
-prettySimpleType _        prty (SimpleType TList1    s1) = [di|[#{prty s1}]|]
-prettySimpleType _        prty (SimpleType TSet0     s1) = [di|[#{prty s1}]|]
-prettySimpleType _        prty (SimpleType TSet1     s1) = [di|[#{prty s1}]|]
-prettySimpleType _       _prty (InlineEnum pt1       s1) =
-  [di|\# InlineEnum unsupported: #{pt1} #{parens (pretty $ PT2 s1)}|]
-
-prettyMaybeType :: String -> (T.Text -> Doc ann) -> Maybe TypeSig -> Doc ann
-prettyMaybeType _ _inner Nothing   = ""
-prettyMaybeType t inner (Just ts) = [di|: #{prettySimpleType t inner ts}|]
 
 -- | comment a block of lines
 commentWith :: T.Text -> [T.Text] -> Doc ann
