+++ lib/haskell/natural4/src/LS/Rule.hs
@@ -9,12 +9,9 @@ module LS.Rule
   ( Rule (..),
     RuleBody (..),
     RuleLabel,
-    Interpreted (..),
     Parser,
     ValuePredicate (..),
     Expect (..),
-    RuleGraph,
-    RuleGraphEdgeLabel,
     _Constitutive,
     _DefNameAlias,
     _DefTypically,
@@ -27,7 +24,6 @@ module LS.Rule
     _RuleGroup,
     _TypeDecl,
     defaultHorn,
-    defaultL4I,
     defaultReg,
     defaultTypeDecl,
     defaultValuePredicate,
@@ -65,8 +61,7 @@ import Control.Monad.Writer.Lazy (WriterT (runWriterT))
 import Data.Aeson (ToJSON)
 import Data.Bifunctor (second)
 import Data.Generics.Product.Types (HasTypes, types)
-import Data.Graph.Inductive (Gr, empty)
-import Data.HashMap.Strict qualified as Map
+import Data.Graph.Inductive (Gr)
 import Data.Hashable (Hashable)
 import Data.List.NonEmpty (NonEmpty(..))
 import Data.Set qualified as Set
@@ -77,7 +72,6 @@ import GHC.Generics (Generic)
 import LS.Types
   ( BoolStructP,
     BoolStructR,
-    ClsTab (..),
     DList,
     Deontic (DMust),
     Depth,
@@ -98,7 +92,6 @@ import LS.Types
     RelationalPredicate (RPMT, RPParamText),
     RuleName,
     RunConfig (debug, parseCallStack),
-    ScopeTabs,
     SrcRef (..),
     TemporalConstraint,
     TypeSig(..),
@@ -113,7 +106,6 @@ import LS.Types
     multiterm2pt,
     rpHead,
   )
-import LS.XPile.Logging (XPileLogW)
 import Optics
   ( Field1 (_1),
     Field2 (_2),
@@ -142,7 +134,7 @@ import Text.Megaparsec
 -- >>> import Optics qualified as O (has)
 -- >>> import Data.Generics.Sum.Constructors
 
-{- | 
+{- |
 [TODO] refactoring: these should be broken out into their own (new)types and have Rule include them all.
 We should take advantage of NoFieldSelectors to reduce the hazards here
 
@@ -155,7 +147,7 @@ True
 
 as well as extracting givens from a rule in an easy way (see the Logical English code).
 
-See https://hackage.haskell.org/package/generic-optics-2.2.1.0/docs/Data-Generics-Sum-Constructors.html 
+See https://hackage.haskell.org/package/generic-optics-2.2.1.0/docs/Data-Generics-Sum-Constructors.html
 for an explanation of how the Generics and optics stuff works
 -}
 data Rule = Regulative
@@ -492,52 +484,6 @@ data Expect = ExpRP      RelationalPredicate
 
 makePrisms ''Rule
 
--- | This is generated by the Interpreter and handed around to different transpilers.
---
--- In the future we may add to this list of attributes the following frequently called functions:
---
--- * `LS.Interpreter.qaHornsT` -- currently a function run against `l4i`
---
--- * `LS.Interpreter.qaHornsR` -- currently a function run against `l4i`
-
-data Interpreted = L4I {
-  -- | all the `DECLARE`d classes
-  classtable :: ClsTab
-
-  -- | all the rule scopes; the exact semantics are not
-  -- entirely nailed down as we don't have a rigorous notion of scope
-  -- yet. This is intended to reflect the real-world usage of "for the
-  -- purposes of this section, X means Y". We might think of that as a
-  -- nested block scope for @let@ bindings.
-  , scopetable :: ScopeTabs
-
-  -- | the original rules on which the interpretation was based. You
-  -- may sometimes see functions that take both 'Interpreted' and
-  -- @[Rule]@; the latter is technically redundant and can be safely
-  -- eliminated. [TODO].
-  , origrules  :: [Rule]
-
-  -- | valuepredicates contain the bulk of the top-level decision logic, and can be easily expressed as instance or class methosd.
-  , valuePreds :: [ValuePredicate]
-
-  -- | rule decision graph gets used by multiple transpilers, so it lives here
-  , ruleGraph :: RuleGraph
-  , ruleGraphErr :: XPileLogW
-  }
-  deriving (Eq, Show)
-
--- | default L4I
-defaultL4I :: Interpreted
-defaultL4I = L4I
-  { classtable = CT Map.empty
-  , scopetable = Map.empty
-  , origrules = mempty
-  , valuePreds = mempty
-  , ruleGraph = empty
-  , ruleGraphErr = mempty
-  }
-
-
 -- | when the input says @DECIDE ClassA's RecordAttr's AttributeNAME IS foo WHEN bar@
 -- we rewrite that to a `ValuePredicate`.
 data ValuePredicate = ValPred
@@ -546,7 +492,7 @@ data ValuePredicate = ValPred
   , objPath    :: [EntityName]  -- ClassA, ClassB, RecordAttrName --> ClassA.ClassB
                   -- If this list is null, then the "attribute" is toplevel / module-global
   , attrName   ::  EntityName   -- ClassA, ClassB, RecordAttrName --> RecordAttrName
-  , attrRel    ::  Maybe RPRel  -- 
+  , attrRel    ::  Maybe RPRel  --
   , attrVal    ::  Maybe RelationalPredicate
   , attrCond   ::  Maybe BoolStructR
   , attrIType  ::  Inferrable TypeSig
@@ -571,18 +517,6 @@ defaultValuePredicate = ValPred
   , origRule   = Nothing
   }
 
--- | structure the rules as a graph.
--- in the simple case, the graph is one or more trees, each rooted at a "top-level" rule which is not "used" by any another rule.
---
--- if we walk the roots, we will sooner or later encounter all the decision elements relevant to each root.
--- in a less simple case, the graph is cyclic! everything depends on everything else! but we can recognize that as an error condition.
---
--- note that a regulative rule R1 HENCE R2 is recorded as a single rule, even if we think of the R2 as a separate rule
--- perhaps we should have a notion of anonymous rules, that are internally labelled and structured, so R2 is equal to R1 in the graph.
-
-type RuleGraphEdgeLabel = ()
-type RuleGraph = Gr Rule RuleGraphEdgeLabel
-
 multiterm2bsr :: Rule -> BoolStructR
 multiterm2bsr = AA.mkLeaf . RPParamText . multiterm2pt . name
 
@@ -598,7 +532,7 @@ pYLocation = snd pXYLocation
 pXYLocation :: (Parser Depth, Parser Depth)
 pXYLocation = (go testx "x location", go testy "y location")
   where
-    go test txt = token (pure . unPos . test) Set.empty <|> pure 0 <?> txt 
+    go test txt = token (pure . unPos . test) Set.empty <|> pure 0 <?> txt
     testx WithPos {pos= SourcePos _ _y x} = x
     testy WithPos {pos= SourcePos _ y _x} = y
