+++ lib/haskell/natural4/src/LS/Verdict.hs
@@ -18,7 +18,7 @@ where
 import AnyAll qualified as AA
 -- import Explainable
 -- import Explainable.MathLang
-import LS.Rule (Interpreted)
+import LS.Interpreter (Interpreted)
 
 -- | User input takes the form of a Marking -- see AnyAll for more on what that means.
 type RuntimeInput = AA.TextMarking
