+++ lib/haskell/natural4/src/LS/XPile/IntroShoehorn.hs
@@ -11,9 +11,8 @@ import Control.Monad.Identity (Identity)
 import Control.Monad.RWS (RWST, ask, evalRWS, tell)
 import Data.Bifunctor (first)
 import Data.Text qualified as Text
-import LS.Interpreter (qaHornsT)
+import LS.Interpreter (Interpreted (..), qaHornsT)
 import LS.PrettyPrinter (myrender, (<//>), (</>))
-import LS.Rule (Interpreted (..))
 import LS.XPile.IntroReader (MyEnv (..))
 import LS.XPile.Logging (XPileLogS, XPileLogW)
 import Prettyprinter (Doc, pretty)
